requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.1
django-extra-views==0.13.0
djangorestframework==3.12.2
Markdown==3.3.4
Pillow==8.1.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
